+++ apps/rig_inbound_gateway/config/prod.exs
@@ -15,9 +15,6 @@ config :rig, RigInboundGatewayWeb.Endpoint,
   env: :prod,
   check_origin: false
 
-# Do not print debug messages in production
-config :logger, level: :info
-
 # ## Using releases
 #
 # If you are doing OTP releases, you need to instruct Phoenix